+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # mei-friend-online CHANGELOG.md
 
+### 1.0.9 patch
+* Facsimile panel GUI elements update local storage 
+* Verovio settings retrieved from local storage
+
+### 1.0.8 patch (released on 21 Jan 2024)
+* Handle facsimile graphics path correctly in GitHub repos with subdirectories
+* Optimize facsimile redrawing after editing the encoding 
+* Polish facsimile updating and size handling
+
 ### 1.0.7 patch (released on 16 Jan 2024)
 * Fix for issue #76 removing `SHIFT + R` as keyboard shortcut for testing measure renumbering, but allowing to enter an upper-case `R`.
 * Small language fixes in DE, PO, and EO.